Trying To Improve The Performance Using Separate Layers
This Commit tries to fix the issue of re-rendering the stuff which would be not changing that often, it uses 2 different ImGui Contexts to draw stuff on 2 different places.

The more interactable stuff is placed in the "mainCtx" which is stuff like buttons and color palette, and more static type of stuff like text is in the "secondaryCtx".